+++ utils/cashback.go
@@ -2,6 +2,6 @@ package utils
 
 import "backend/env"
 
-func ApplyCashbackRate(value float32) float32 {
+func ApplyCashbackRate(value int) int {
   return (env.CashbackRate * value) / 100
 }
